moved token logic to azureController
